+++ filein.txt
@@ -1,8 +1,6 @@
-2
-12 18
-2
-24 5
+3
+15 36
 1
-12 169
-1
-12 -169
+4 5
+2
+0 3
